+++ packages/node/test/parsers.test.ts
@@ -1,7 +1,7 @@
 import * as fs from 'fs';
 
 import * as Parsers from '../src/parsers';
-import * as stacktrace from '../src/stack-trace';
+import * as stacktrace from '../src/stacktrace';
 
 import { getError } from './helper/error';